Auto stash before merge of "master" and "origin/master"

+++ ActividadesProg2020/src/AA2_01/SceneGame.h
@@ -25,7 +25,7 @@ private:
 	//Map
 	
 	clock_t lastTime;
-	int timeDown, playTime;
+	float timeDown, playTime;
 	float deltaTime;
 
 public:
